Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int a;
-    int b;
-    int c;
+    float a;
+    float b;
+    float c;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
